requirements.txt
pypinyin>=0.46.0
PyYAML>=6.0
unidecode>=1.3.6
